audit-cli/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentcollect-audit"
version = "0.1.0"
description = "AgentCollect Audit CLI — Automated AR diagnostic reports for B2B prospects"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "AgentCollect", email = "founders@agentcollect.com" }]
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "rich>=13.7",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "pydantic>=2.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
agentcollect-audit = "agentcollect_audit.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"agentcollect_audit" = ["templates/*.j2", "data/*.json"]

[tool.ruff]
line-length = 100
target-version = "py311"
